Enable production builds with offline template compilation
Since this is Electron we don't have to minify but bundling
helps a lot. The offline template compiler had issues with
the UMD bundles that angular distributes so I had to switch
to the plain JS files. Unfortunately they are ES5 so tree-shaking
is not possible when bundling.

For bundling I use SystemJS builder, but unfortunately it has some
issues when dealing with folders outside the baseUrl, so I had to
work around this by switching the CWD to the baseURL path before
requiring the Builder.